requirements.txt
streamlit==1.31.1
qdrant-client==1.7.3
transformers==4.37.2
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
groq==0.9.0
httpx==0.27.0
python-dotenv==1.0.1
sentence-transformers==2.3.1
python-dateutil==2.8.2
